FIR/UAST: commonize PsiElement conversion
