+++ uast/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/FirKotlinConverter.kt
@@ -5,21 +5,17 @@
 
 package org.jetbrains.uast.kotlin
 
-import com.intellij.psi.PsiComment
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiEnumConstant
 import com.intellij.psi.PsiField
-import com.intellij.psi.impl.source.tree.LeafPsiElement
 import org.jetbrains.kotlin.asJava.LightClassUtil
 import org.jetbrains.kotlin.asJava.classes.KtLightClass
 import org.jetbrains.kotlin.asJava.elements.*
 import org.jetbrains.kotlin.asJava.findFacadeClass
 import org.jetbrains.kotlin.asJava.toLightClass
 import org.jetbrains.kotlin.asJava.toPsiParameters
-import org.jetbrains.kotlin.kdoc.psi.impl.KDocName
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.*
-import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
 import org.jetbrains.kotlin.utils.addToStdlib.safeAs
 import org.jetbrains.uast.*
 import org.jetbrains.uast.expressions.UInjectionHost
@@ -32,19 +28,6 @@ import org.jetbrains.uast.kotlin.psi.UastKotlinPsiParameterBase
 import org.jetbrains.uast.kotlin.psi.UastKotlinPsiVariable
 
 internal object FirKotlinConverter : BaseKotlinConverter {
-
-    internal fun convertDeclarationOrElement(
-        element: PsiElement,
-        givenParent: UElement?,
-        requiredTypes: Array<out Class<out UElement>>
-    ): UElement? {
-        if (element is UElement) return element
-
-        // TODO: cache inside PsiElement?
-
-        return convertDeclaration(element, givenParent, requiredTypes)
-            ?: convertPsiElement(element, givenParent, requiredTypes)
-    }
 
     override fun convertDeclaration(
         element: PsiElement,
@@ -191,45 +174,6 @@ internal object FirKotlinConverter : BaseKotlinConverter {
         )
     }
 
-    internal fun convertClassOrObject(
-        element: KtClassOrObject,
-        givenParent: UElement?,
-        requiredTypes: Array<out Class<out UElement>>
-    ): Sequence<UElement> {
-        val ktLightClass = element.toLightClass() ?: return emptySequence()
-        val uClass = KotlinUClass.create(ktLightClass, givenParent)
-        return requiredTypes.accommodate(
-            // Class
-            alternative { uClass },
-            // Object
-            alternative primaryConstructor@{
-                val primaryConstructor = element.primaryConstructor ?: return@primaryConstructor null
-                uClass.methods.asSequence()
-                    .filter { it.sourcePsi == primaryConstructor }
-                    .firstOrNull()
-            }
-        )
-    }
-
-    private fun convertToPropertyAlternatives(
-        methods: LightClassUtil.PropertyAccessorsPsiMethods?,
-        givenParent: UElement?
-    ): Array<UElementAlternative<*>> =
-        if (methods != null)
-            arrayOf(
-                alternative { methods.backingField?.let { KotlinUField(it, getKotlinMemberOrigin(it), givenParent) } },
-                alternative { methods.getter?.let { convertDeclaration(it, givenParent, arrayOf(UMethod::class.java)) as? UMethod } },
-                alternative { methods.setter?.let { convertDeclaration(it, givenParent, arrayOf(UMethod::class.java)) as? UMethod } }
-            )
-        else emptyArray()
-
-    internal fun convertNonLocalProperty(
-        property: KtProperty,
-        givenParent: UElement?,
-        requiredTypes: Array<out Class<out UElement>>
-    ): Sequence<UElement> =
-        requiredTypes.accommodate(*convertToPropertyAlternatives(LightClassUtil.getLightClassPropertyMethods(property), givenParent))
-
     internal fun convertParameter(
         element: KtParameter,
         givenParent: UElement?,
@@ -246,110 +190,6 @@ internal object FirKotlinConverter : BaseKotlinConverter {
             },
             *convertToPropertyAlternatives(LightClassUtil.getLightClassPropertyMethods(element), givenParent)
         )
-
-    override fun convertPsiElement(
-        element: PsiElement?,
-        givenParent: UElement?,
-        requiredTypes: Array<out Class<out UElement>>
-    ): UElement? {
-        if (element == null) return null
-
-        fun <P : PsiElement> build(ctor: (P, UElement?) -> UElement): () -> UElement? = {
-            @Suppress("UNCHECKED_CAST")
-            ctor(element as P, givenParent)
-        }
-
-        return with(requiredTypes) {
-            when (element) {
-                is KtParameterList -> {
-                    el<UDeclarationsExpression> {
-                        val declarationsExpression = KotlinUDeclarationsExpression(null, givenParent, null)
-                        declarationsExpression.apply {
-                            declarations = element.parameters.mapIndexed { i, p ->
-                                KotlinUParameter(UastKotlinPsiParameter.create(p, element, declarationsExpression, i), p, this)
-                            }
-                        }
-                    }
-                }
-                is KtClassBody -> {
-                    el<UExpressionList>(build(KotlinUExpressionList.Companion::createClassBody))
-                }
-                is KtCatchClause -> {
-                    el<UCatchClause>(build(::KotlinUCatchClause))
-                }
-                is KtVariableDeclaration -> {
-                    if (element is KtProperty && !element.isLocal) {
-                        convertNonLocalProperty(element, givenParent, this).firstOrNull()
-                    } else {
-                        el<UVariable> { convertVariablesDeclaration(element, givenParent).declarations.singleOrNull() }
-                            ?: expr<UDeclarationsExpression> { convertExpression(element, givenParent, requiredTypes) }
-                    }
-                }
-                is KtExpression -> {
-                    convertExpression(element, givenParent, requiredTypes)
-                }
-                is KtLiteralStringTemplateEntry, is KtEscapeStringTemplateEntry -> {
-                    el<ULiteralExpression>(build(::KotlinStringULiteralExpression))
-                }
-                is KtStringTemplateEntry -> {
-                    element.expression?.let { convertExpression(it, givenParent, requiredTypes) }
-                        ?: expr<UExpression> { UastEmptyExpression(givenParent) }
-                }
-                is KtWhenEntry -> el<USwitchClauseExpressionWithBody>(build(::KotlinUSwitchEntry))
-                is KtWhenCondition -> convertWhenCondition(element, givenParent, requiredTypes)
-                is KtTypeReference ->
-                    requiredTypes.accommodate(
-                        alternative { KotlinUTypeReferenceExpression(element, givenParent) },
-                        alternative { convertReceiverParameter(element) }
-                    ).firstOrNull()
-                is KtConstructorDelegationCall ->
-                    el<UCallExpression> { KotlinUFunctionCallExpression(element, givenParent) }
-                is KtSuperTypeCallEntry ->
-                    el<UExpression> {
-                        (element.getParentOfType<KtClassOrObject>(true)?.parent as? KtObjectLiteralExpression)
-                            ?.toUElementOfType<UExpression>()
-                            ?: KotlinUFunctionCallExpression(element, givenParent)
-                    }
-                is KtImportDirective -> {
-                    el<UImportStatement>(build(::KotlinUImportStatement))
-                }
-                is PsiComment -> el<UComment>(build(::UComment))
-                is KDocName -> {
-                    if (element.getQualifier() == null)
-                        el<USimpleNameReferenceExpression> {
-                            element.lastChild?.let { psiIdentifier ->
-                                KotlinStringUSimpleReferenceExpression(psiIdentifier.text, givenParent, element, element)
-                            }
-                        }
-                    else el<UQualifiedReferenceExpression>(build(::KotlinDocUQualifiedReferenceExpression))
-                }
-                is LeafPsiElement -> {
-                    when {
-                        element.elementType in identifiersTokens -> {
-                            if (element.elementType != KtTokens.OBJECT_KEYWORD ||
-                                element.getParentOfType<KtObjectDeclaration>(false)?.nameIdentifier == null
-                            )
-                                el<UIdentifier>(build(::KotlinUIdentifier))
-                            else null
-                        }
-                        element.elementType in KtTokens.OPERATIONS && element.parent is KtOperationReferenceExpression -> {
-                            el<UIdentifier>(build(::KotlinUIdentifier))
-                        }
-                        element.elementType == KtTokens.LBRACKET && element.parent is KtCollectionLiteralExpression -> {
-                            el<UIdentifier> {
-                                UIdentifier(
-                                    element,
-                                    KotlinUCollectionLiteralExpression(element.parent as KtCollectionLiteralExpression, null)
-                                )
-                            }
-                        }
-                        else -> null
-                    }
-                }
-                else -> null
-            }
-        }
-    }
 
     // TODO: forceUInjectionHost (for test)?
