+++ uast/uast-kotlin/src/org/jetbrains/uast/kotlin/KotlinConverter.kt
@@ -7,17 +7,14 @@ package org.jetbrains.uast.kotlin
 
 import com.intellij.openapi.util.registry.Registry
 import com.intellij.psi.*
-import com.intellij.psi.impl.source.tree.LeafPsiElement
 import org.jetbrains.annotations.TestOnly
 import org.jetbrains.kotlin.asJava.LightClassUtil
 import org.jetbrains.kotlin.asJava.classes.KtLightClass
 import org.jetbrains.kotlin.asJava.elements.*
 import org.jetbrains.kotlin.asJava.findFacadeClass
 import org.jetbrains.kotlin.asJava.toLightClass
-import org.jetbrains.kotlin.kdoc.psi.impl.KDocName
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.*
-import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
 import org.jetbrains.kotlin.utils.addToStdlib.safeAs
 import org.jetbrains.uast.*
 import org.jetbrains.uast.expressions.UInjectionHost
@@ -25,106 +22,6 @@ import org.jetbrains.uast.kotlin.expressions.*
 import org.jetbrains.uast.kotlin.psi.*
 
 internal object KotlinConverter : BaseKotlinConverter {
-
-    override fun convertPsiElement(
-        element: PsiElement?,
-        givenParent: UElement?,
-        requiredTypes: Array<out Class<out UElement>>
-    ): UElement? {
-        if (element == null) return null
-
-        fun <P : PsiElement> build(ctor: (P, UElement?) -> UElement): () -> UElement? {
-            return {
-                @Suppress("UNCHECKED_CAST")
-                ctor(element as P, givenParent)
-            }
-        }
-
-        return with (requiredTypes) { when (element) {
-            is KtParameterList -> el<UDeclarationsExpression> {
-                val declarationsExpression = KotlinUDeclarationsExpression(null, givenParent, null)
-                declarationsExpression.apply {
-                    declarations = element.parameters.mapIndexed { i, p ->
-                        KotlinUParameter(UastKotlinPsiParameter.create(p, element, declarationsExpression, i), p, this)
-                    }
-                }
-            }
-            is KtClassBody -> el<UExpressionList>(build(KotlinUExpressionList.Companion::createClassBody))
-            is KtCatchClause -> el<UCatchClause>(build(::KotlinUCatchClause))
-            is KtVariableDeclaration ->
-                if (element is KtProperty && !element.isLocal) {
-                    convertNonLocalProperty(element, givenParent, this).firstOrNull()
-                }
-                else {
-                    el<UVariable> { convertVariablesDeclaration(element, givenParent).declarations.singleOrNull() }
-                        ?: expr<UDeclarationsExpression> { convertExpression(element, givenParent, requiredTypes) }
-                }
-
-            is KtExpression -> convertExpression(element, givenParent, requiredTypes)
-            is KtLambdaArgument -> element.getLambdaExpression()?.let { convertExpression(it, givenParent, requiredTypes) }
-            is KtLightElementBase -> {
-                val expression = element.kotlinOrigin
-                when (expression) {
-                    is KtExpression -> convertExpression(expression, givenParent, requiredTypes)
-                    else -> el<UExpression> { UastEmptyExpression(givenParent) }
-                }
-            }
-            is KtLiteralStringTemplateEntry, is KtEscapeStringTemplateEntry ->
-                el<ULiteralExpression>(build(::KotlinStringULiteralExpression))
-            is KtStringTemplateEntry ->
-                element.expression?.let { convertExpression(it, givenParent, requiredTypes) }
-                    ?: expr<UExpression> { UastEmptyExpression(givenParent) }
-            is KtWhenEntry -> el<USwitchClauseExpressionWithBody>(build(::KotlinUSwitchEntry))
-            is KtWhenCondition -> convertWhenCondition(element, givenParent, requiredTypes)
-            is KtTypeReference ->
-                requiredTypes.accommodate(
-                    alternative { KotlinUTypeReferenceExpression(element, givenParent) },
-                    alternative { convertReceiverParameter(element) }
-                ).firstOrNull()
-            is KtConstructorDelegationCall ->
-                el<UCallExpression> { KotlinUFunctionCallExpression(element, givenParent) }
-            is KtSuperTypeCallEntry ->
-                el<UExpression> {
-                    (element.getParentOfType<KtClassOrObject>(true)?.parent as? KtObjectLiteralExpression)
-                        ?.toUElementOfType<UExpression>()
-                        ?: KotlinUFunctionCallExpression(element, givenParent)
-                }
-            is KtImportDirective -> el<UImportStatement>(build(::KotlinUImportStatement))
-            is PsiComment -> el<UComment>(build(::UComment))
-            is KDocName -> {
-                if (element.getQualifier() == null)
-                    el<USimpleNameReferenceExpression> {
-                        element.lastChild?.let { psiIdentifier ->
-                            KotlinStringUSimpleReferenceExpression(psiIdentifier.text, givenParent, element, element)
-                        }
-                    }
-                else el<UQualifiedReferenceExpression>(build(::KotlinDocUQualifiedReferenceExpression))
-            }
-            else -> {
-                if (element is LeafPsiElement) {
-                    if (element.elementType in identifiersTokens)
-                        if (element.elementType != KtTokens.OBJECT_KEYWORD ||
-                            element.getParentOfType<KtObjectDeclaration>(false)?.nameIdentifier == null
-                        )
-                            el<UIdentifier>(build(::KotlinUIdentifier))
-                        else null
-                    else if (element.elementType in KtTokens.OPERATIONS && element.parent is KtOperationReferenceExpression)
-                        el<UIdentifier>(build(::KotlinUIdentifier))
-                    else if (element.elementType == KtTokens.LBRACKET && element.parent is KtCollectionLiteralExpression)
-                        el<UIdentifier> {
-                            UIdentifier(
-                                element,
-                                KotlinUCollectionLiteralExpression(
-                                    element.parent as KtCollectionLiteralExpression,
-                                    null
-                                )
-                            )
-                        }
-                    else null
-                } else null
-            }
-        }}
-    }
 
     var forceUInjectionHost = Registry.`is`("kotlin.uast.force.uinjectionhost", false)
         @TestOnly
@@ -389,33 +286,6 @@ internal object KotlinConverter : BaseKotlinConverter {
         return getContainingLightClass(original)
     }
 
-    fun convertDeclarationOrElement(
-        element: PsiElement,
-        givenParent: UElement?,
-        expectedTypes: Array<out Class<out UElement>>
-    ): UElement? {
-        return if (element is UElement) element
-        else convertDeclaration(element, givenParent, expectedTypes)
-            ?: KotlinConverter.convertPsiElement(element, givenParent, expectedTypes)
-    }
-
-    private fun convertToPropertyAlternatives(
-        methods: LightClassUtil.PropertyAccessorsPsiMethods?,
-        givenParent: UElement?,
-    ): Array<UElementAlternative<*>> = if (methods != null) arrayOf(
-        alternative { methods.backingField?.let { KotlinUField(it, (it as? KtLightElement<*, *>)?.kotlinOrigin, givenParent) } },
-        alternative { methods.getter?.let { convertDeclaration(it, givenParent, arrayOf(UMethod::class.java)) as? UMethod } },
-        alternative { methods.setter?.let { convertDeclaration(it, givenParent, arrayOf(UMethod::class.java)) as? UMethod } }
-    ) else emptyArray()
-
-    fun convertNonLocalProperty(
-        property: KtProperty,
-        givenParent: UElement?,
-        expectedTypes: Array<out Class<out UElement>>
-    ): Sequence<UElement> =
-        expectedTypes.accommodate(*convertToPropertyAlternatives(LightClassUtil.getLightClassPropertyMethods(property), givenParent))
-
-
     fun convertParameter(
         element: KtParameter,
         givenParent: UElement?,
@@ -439,24 +309,6 @@ internal object KotlinConverter : BaseKotlinConverter {
         },
         *convertToPropertyAlternatives(LightClassUtil.getLightClassPropertyMethods(element), givenParent)
     )
-
-    fun convertClassOrObject(
-        element: KtClassOrObject,
-        givenParent: UElement?,
-        expectedTypes: Array<out Class<out UElement>>
-    ): Sequence<UElement> {
-        val ktLightClass = element.toLightClass() ?: return emptySequence()
-        val uClass = KotlinUClass.create(ktLightClass, givenParent)
-        return expectedTypes.accommodate(
-            alternative { uClass },
-            alternative primaryConstructor@{
-                val primaryConstructor = element.primaryConstructor ?: return@primaryConstructor null
-                uClass.methods.asSequence()
-                    .filter { it.sourcePsi == primaryConstructor }
-                    .firstOrNull()
-            }
-        )
-    }
 
     fun convertKtFile(
         element: KtFile,
